put types from lib rs into types file
